+++ packages/file_selector/file_selector_android/CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial Android implementation of `file_selector`.